# internal/handlers/chat_handler.go
package handlers

import (
	"empre_backend/internal/services"
	"empre_backend/internal/websocket"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

type ChatHandler struct {
	Hub     *websocket.Hub
	service *services.ChatService
}

func NewChatHandler(hub *websocket.Hub, service *services.ChatService) *ChatHandler {
	return &ChatHandler{
		Hub:     hub,
		service: service,
	}
}

// HandleWebSocket initiates a real-time chat connection
// @Summary WebSocket Chat
// @Description Upgrade to WebSocket for real-time messaging
// @Tags Chat
// @Security BearerAuth
// @Param token query string true "JWT Token"
// @Router /api/chat/ws [get]
func (h *ChatHandler) HandleWebSocket(c *gin.Context) {
	// Get User ID from context (set by auth middleware)
	userIDVal, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "User not authenticated"})
		return
	}
	userID := userIDVal.(uuid.UUID)

	websocket.ServeWs(h.Hub, c, userID)
}

// FindAllConversations retrieves all active conversations for the user
// @Summary List conversations
// @Description Get a list of the last message from each active conversation
// @Tags Chat
// @Produce json
// @Security BearerAuth
// @Success 200 {array} models.Message
// @Failure 500 {object} map[string]string
// @Router /api/chat/conversations [get]
func (h *ChatHandler) FindAllConversations(c *gin.Context) {
	userIDVal, _ := c.Get("userID")
	userID := userIDVal.(uuid.UUID)

	conversations, err := h.service.FindAllConversations(userID)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, conversations)
}

// FindMessagesHistory retrieves full message history between a user and an entity
// @Summary Message history
// @Description Get all messages in a conversation with a specific business
// @Tags Chat
// @Produce json
// @Security BearerAuth
// @Param entity_id path string true "Entity ID"
// @Param user_id query string false "User ID (Owner only usage)"
// @Success 200 {array} models.Message
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/chat/history/{entity_id} [get]
func (h *ChatHandler) FindMessagesHistory(c *gin.Context) {
	entityIDStr := c.Param("entity_id")
	userIDStr := c.Query("user_id")

	entityID, err := uuid.Parse(entityIDStr)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid Entity ID format"})
		return
	}

	// Current authenticated user (can be customer or owner)
	currentUserIDVal, _ := c.Get("userID")
	currentUserID := currentUserIDVal.(uuid.UUID)

	var targetUserID uuid.UUID
	if userIDStr != "" {
		// If an owner is requesting history with a specific customer
		targetUserID, err = uuid.Parse(userIDStr)
		if err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid User ID format"})
			return
		}
	} else {
		// If a customer is requesting their own history with an entity
		targetUserID = currentUserID
	}

	messages, err := h.service.FindMessagesHistory(entityID, targetUserID)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, messages)
}